refactor(engine): collapse QueryError into StructuredQueryError
Merge the engine-internal QueryError type into StructuredQueryError,
eliminating one intermediate type from the error chain. Constructors
(missing_required_filter, http_provider_request) now live directly on
StructuredQueryError. The struct definition, Display impl, getters,
and constructors all live in query_error.rs; error.rs re-exports the
type.

The conversion in provider_error_to_core now builds
StructuredQueryError directly without an intermediate step.
